+++ datasette/templates/table.html
@@ -9,6 +9,7 @@
 {% if table_page_data.foreignKeys %}
 <script src="{{ urls.static('autocomplete.js') }}" defer></script>
 {% endif %}
+<script src="{{ urls.static('edit-tools.js') }}?hash={{ edit_tools_js_hash }}" defer></script>
 <script src="{{ urls.static('table.js') }}?hash={{ table_js_hash }}" defer></script>
 <script src="{{ urls.static('mobile-column-actions.js') }}" defer></script>
 <script>DATASETTE_ALLOW_FACET = {{ datasette_allow_facet }};</script>
